cleaning up code , and start integartion with canvas
